:shirt: Refactor perftests and fix warnings
