+++ src/Directory.Packages.props
@@ -11,9 +11,11 @@
         <PackageVersion Include="NUnit3TestAdapter" Version="4.2.1" />
         <PackageVersion Include="Microsoft.NET.Test.Sdk" Version="17.1.0"/>
         <PackageVersion Include="coverlet.collector" Version="3.1.2" />
-        <PackageVersion Include="BenchmarkDotNet" Version="0.13.1" />
         <PackageVersion Include="Moq" Version="4.16.1" />
 
+        <PackageVersion Include="BenchmarkDotNet" Version="0.13.1" />
+        <PackageVersion Include="System.CommandLine" Version="2.0.0-beta3.22114.1" />
+        
         <PackageVersion Include="Microsoft.SourceLink.GitHub" Version="1.1.1" />
 
         <PackageVersion Include="StyleCop.Analyzers" Version="1.1.118" />
