+++ src/Yarhl.PerformanceTests/Encodings/index-jis0208.txt
