+++ src/Yarhl.PerformanceTests/Encodings/japanese.txt
