+++ src/Yarhl.PerformanceTests/IO/DataStreamReadWriteTests.cs
@@ -17,7 +17,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
-namespace Yarhl.PerformanceTests
+namespace Yarhl.PerformanceTests.IO
 {
     using System;
     using System.IO;
